Namespace refactorings
