+++ TruRating.TruModule.V2xx.Tests.Unit/TruRating.TruModule.V2xx.Tests.Unit.csproj
@@ -80,15 +80,15 @@
     <Compile Include="Security\MacSignatureCalculatorTests\WhenTestingCalculatorWithInvalidTransportKey.cs" />
     <Compile Include="Serialization\DefaultSerializerTests\DefaultSerializerTests.cs" />
     <Compile Include="Messages\TruServiceMessageFactoryTests.cs" />
+    <Compile Include="TruModuleStandaloneTests\Activate\Foo.cs" />
     <Compile Include="TruModuleStandaloneTests\ActiveTruModuleStandaloneTestsContext.cs" />
-    <Compile Include="TruModuleStandaloneTests\LookupTests\WhenGettingLookupDataAndResoponseDoesNotContainDeviceLangauge.cs" />
-    <Compile Include="TruModuleStandaloneTests\LookupTests\TruModuleStandaloneLookupTestContext.cs" />
-    <Compile Include="TruModuleStandaloneTests\WhenActivating.cs" />
+    <Compile Include="TruModuleStandaloneTests\Lookup\WhenGettingLookupDataAndResoponseDoesNotContainDeviceLangauge.cs" />
+    <Compile Include="TruModuleStandaloneTests\Lookup\TruModuleStandaloneLookupTestContext.cs" />
     <Compile Include="TruModuleStandaloneTests\WhenDoingARatingAndNotActive.cs" />
     <Compile Include="TruModuleStandaloneTests\WhenDoingARatingAndActive.cs" />
-    <Compile Include="TruModuleStandaloneTests\LookupTests\WhenGettingLookupDataAndResoponseContainsDeviceLangauge.cs" />
-    <Compile Include="TruModuleStandaloneTests\WhenSendingATransactionAndActive.cs" />
-    <Compile Include="TruModuleStandaloneTests\WhenSendingATransactionAndNotActive.cs" />
+    <Compile Include="TruModuleStandaloneTests\Lookup\WhenGettingLookupDataAndResoponseContainsDeviceLangauge.cs" />
+    <Compile Include="TruModuleStandaloneTests\SendTransaction\WhenSendingATransactionAndActive.cs" />
+    <Compile Include="TruModuleStandaloneTests\SendTransaction\WhenSendingATransactionAndNotActive.cs" />
     <Compile Include="TruModuleTests\WhenQuestionAvailableAndDoingDwellTimeExtendRating.cs" />
     <Compile Include="TruModuleTests\TruModuleTestsContext.cs" />
     <Compile Include="TruModuleTests\Activation\WhenTruModuleIsInactiveAndTtlNotExpiredAndDoingAPassiveActivationCheck.cs" />
